NLP/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
_tflow_select=2.2.0=eigen
abseil-cpp=20220623.0=hddbf539_6
absl-py=1.4.0=py310hecd8cb5_0
aiohttp=3.9.3=py310h6c40b1e_0
aiosignal=1.3.1=pypi_0
altair=5.3.0=pypi_0
anyio=4.3.0=pypi_0
appnope=0.1.4=pyhd8ed1ab_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.3.0=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
astunparse=1.6.3=py_0
async-lru=2.0.4=pypi_0
async-timeout=4.0.3=py310hecd8cb5_0
attrs=23.2.0=pypi_0
babel=2.14.0=pypi_0
beautifulsoup4=4.12.3=pypi_0
bertopic=0.16.0=pypi_0
blas=1.0=openblas
bleach=6.1.0=pypi_0
blinker=1.7.0=pypi_0
brotli-python=1.0.9=py310he9d5cce_7
bzip2=1.0.8=h6c40b1e_5
c-ares=1.19.1=h6c40b1e_0
ca-certificates=2024.3.11=hecd8cb5_0
cachetools=5.3.3=pypi_0
certifi=2024.2.2=py310hecd8cb5_0
cffi=1.16.0=py310h6c40b1e_0
chardet=4.0.0=py310hecd8cb5_1003
charset-normalizer=3.3.2=pypi_0
cleantext=1.1.4=pypi_0
click=8.1.7=py310hecd8cb5_0
colorama=0.4.6=pyhd8ed1ab_0
comm=0.2.2=pyhd8ed1ab_0
contourpy=1.2.0=pypi_0
cryptography=42.0.5=py310h30e54ef_0
cycler=0.12.1=pypi_0
cython=0.29.37=pypi_0
dateparser=1.2.0=pypi_0
debugpy=1.8.1=py310h5daac23_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pypi_0
exceptiongroup=1.2.0=pyhd8ed1ab_2
executing=2.0.1=pyhd8ed1ab_0
fastjsonschema=2.19.1=pypi_0
filelock=3.13.3=pypi_0
flatbuffers=22.12.06=hf0c8a7f_2
fonttools=4.50.0=pypi_0
fqdn=1.5.1=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2024.3.1=pypi_0
ftfy=6.2.0=pypi_0
gast=0.4.0=pyhd3eb1b0_0
giflib=5.2.1=h6c40b1e_3
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
google-auth=2.22.0=py310hecd8cb5_0
google-auth-oauthlib=0.4.4=pyhd3eb1b0_0
google-pasta=0.2.0=pyhd3eb1b0_0
grpc-cpp=1.51.1=h88f4db0_1
grpcio=1.51.1=py310hdfcfac3_1
h11=0.14.0=pypi_0
h5py=3.9.0=py310hf9d4033_0
hdbscan=0.8.33=pypi_0
hdf5=1.12.1=ha01d115_3
httpcore=1.0.5=pypi_0
httpx=0.27.0=pypi_0
huggingface-hub=0.22.2=pypi_0
icu=70.1=h96cf925_0
idna=3.6=pypi_0
importlib-metadata=7.1.0=pyha770c72_0
importlib_metadata=7.1.0=hd8ed1ab_0
ipykernel=6.29.3=pyh3cd1d5f_0
ipython=8.23.0=pypi_0
ipython-genutils=0.2.0=pypi_0
isoduration=20.11.0=pypi_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.3=pypi_0
joblib=1.4.0=pyhd8ed1ab_0
jpeg=9e=h6c40b1e_1
json5=0.9.24=pypi_0
jsonpointer=2.4=pypi_0
jsonschema=4.21.1=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter-events=0.10.0=pypi_0
jupyter-lsp=2.2.4=pypi_0
jupyter-server=2.13.0=pypi_0
jupyter-server-terminals=0.5.3=pypi_0
jupyter_client=8.6.1=pyhd8ed1ab_0
jupyter_core=5.7.2=py310h2ec42d9_0
jupyterlab=4.1.5=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
jupyterlab-server=2.25.4=pypi_0
keras=2.11.0=py310_0
keras-preprocessing=1.1.2=pyhd3eb1b0_0
kiwisolver=1.4.5=pypi_0
krb5=1.20.1=h428f121_1
libabseil=20220623.0=cxx17_h844d122_6
libblas=3.9.0=22_osx64_openblas
libcblas=3.9.0=22_osx64_openblas
libcurl=8.5.0=hf20ceda_0
libcxx=16.0.6=hd57cbcb_0
libedit=3.1.20230828=h6c40b1e_0
libev=4.33=h9ed2024_1
libffi=3.4.4=hecd8cb5_0
libgfortran=5.0.0=13_2_0_h97931a8_3
libgfortran5=13.2.0=h2873a65_3
libgrpc=1.51.1=h1ddfa78_1
liblapack=3.9.0=22_osx64_openblas
libllvm14=14.0.6=h91fad77_3
libnghttp2=1.57.0=h9beae6a_0
libopenblas=0.3.27=openmp_hfef2a42_0
libpng=1.6.39=h6c40b1e_0
libprotobuf=3.21.12=h7d26f99_2
libsodium=1.0.18=hbcb3906_1
libsqlite=3.45.2=h92b6c6a_0
libssh2=1.10.0=h04015c4_2
libzlib=1.2.13=h8a1eda9_5
llvm-openmp=18.1.3=hb6ac08f_0
llvmlite=0.42.0=py310hcec6c5f_0
markdown=3.4.1=py310hecd8cb5_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.8.3=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mdurl=0.1.2=pypi_0
mistune=3.0.2=pypi_0
mplfinance=0.12.10b0=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.5=pypi_0
nbclient=0.10.0=pypi_0
nbconvert=7.16.3=pypi_0
nbformat=5.10.3=pypi_0
ncurses=6.4=hcec6c5f_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=3.2.1=pypi_0
newsapi-python=0.2.7=pypi_0
nltk=3.8.1=pypi_0
notebook=7.1.2=pypi_0
notebook-shim=0.2.4=pypi_0
numba=0.59.1=py310h1d5af72_0
numpy=1.23.5=py310he50c29a_0
numpy-base=1.23.5=py310h992e150_0
oauthlib=3.2.2=py310hecd8cb5_0
openssl=3.0.13=hca72f7f_0
opt_einsum=3.3.0=pyhd3eb1b0_1
outcome=1.3.0.post0=pypi_0
overrides=7.7.0=pypi_0
packaging=24.0=pyhd8ed1ab_0
pandas=2.2.1=pypi_0
pandocfilters=1.5.1=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.3.0=pypi_0
pip=23.3.1=py310hecd8cb5_0
platformdirs=4.2.0=pyhd8ed1ab_0
plotly=5.20.0=pypi_0
praw=7.7.1=pypi_0
prawcore=2.4.0=pypi_0
prometheus-client=0.20.0=pypi_0
prompt-toolkit=3.0.42=pyha770c72_0
protobuf=4.25.3=pypi_0
psutil=5.9.8=py310hb372a2b_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
py4j=0.10.9.7=pypi_0
pyarrow=15.0.2=pypi_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.22=pypi_0
pycryptodome=3.20.0=pypi_0
pydeck=0.8.1b0=pypi_0
pygments=2.17.2=pyhd8ed1ab_0
pyjwt=2.4.0=py310hecd8cb5_0
pynndescent=0.5.12=pyhca7485f_0
pyopenssl=24.0.0=py310hecd8cb5_0
pyparsing=3.1.2=pypi_0
pysocks=1.7.1=py310hecd8cb5_0
pyspark=3.5.1=pypi_0
python=3.10.14=h5ee71fb_0
python-binance=1.0.19=pypi_0
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
python-flatbuffers=2.0=pyhd3eb1b0_0
python-json-logger=2.0.7=pypi_0
python_abi=3.10=2_cp310
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=25.1.2=py310h6b67f7f_0
ratelimit=2.2.1=pypi_0
re2=2023.02.01=hf0c8a7f_0
readline=8.2=hca72f7f_0
referencing=0.34.0=pypi_0
regex=2023.12.25=pypi_0
requests=2.31.0=py310hecd8cb5_1
requests-oauthlib=1.3.0=py_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rich=13.7.1=pypi_0
rpds-py=0.18.0=pypi_0
rsa=4.7.2=pyhd3eb1b0_1
safetensors=0.4.2=pypi_0
scikit-learn=1.4.1.post1=py310h38ce860_0
scipy=1.13.0=py310hdfaad59_0
seaborn=0.13.2=pypi_0
selenium=4.19.0=pypi_0
send2trash=1.8.2=pypi_0
sentence-transformers=2.6.1=pypi_0
setuptools=68.2.2=py310hecd8cb5_0
six=1.16.0=pyh6c4a22f_0
smmap=5.0.1=pypi_0
snappy=1.1.10=hcec6c5f_1
sniffio=1.3.1=pypi_0
sortedcontainers=2.4.0=pypi_0
soupsieve=2.5=pypi_0
sqlite=3.41.2=h6c40b1e_0
stack_data=0.6.2=pyhd8ed1ab_0
streamlit=1.33.0=pypi_0
sympy=1.12=pypi_0
tbb=2021.10.0=h1c7c39f_2
tenacity=8.2.3=pypi_0
tensorboard=2.11.0=py310_0
tensorboard-data-server=0.6.1=py310h7242b5c_0
tensorboard-plugin-wit=1.6.0=py_0
tensorflow=2.11.0=cpu_py310h22f808f_0
tensorflow-base=2.11.0=cpu_py310h760b059_0
tensorflow-estimator=2.11.0=cpu_py310h5e669bb_0
termcolor=2.1.0=py310hecd8cb5_0
terminado=0.18.1=pypi_0
threadpoolctl=3.4.0=pyhc1e730c_0
tinycss2=1.2.1=pypi_0
tk=8.6.12=h5d9f67b_0
tokenizers=0.15.2=pypi_0
toml=0.10.2=pypi_0
tomli=2.0.1=pypi_0
toolz=0.12.1=pypi_0
torch=2.2.2=pypi_0
tornado=6.4=py310hb372a2b_0
tqdm=4.66.2=pyhd8ed1ab_0
traitlets=5.14.2=pyhd8ed1ab_0
transformers=4.39.2=pypi_0
trio=0.25.0=pypi_0
trio-websocket=0.11.1=pypi_0
types-python-dateutil=2.9.0.20240316=pypi_0
typing_extensions=4.10.0=pyha770c72_0
tzdata=2024.1=pypi_0
tzlocal=5.2=pypi_0
ujson=5.9.0=pypi_0
umap=0.1.1=pypi_0
umap-learn=0.5.6=pypi_0
update-checker=0.18.0=pypi_0
uri-template=1.3.0=pypi_0
urllib3=2.2.1=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
webcolors=1.13=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.7.0=pypi_0
websockets=12.0=pypi_0
werkzeug=2.3.8=py310hecd8cb5_0
wheel=0.35.1=pyhd3eb1b0_0
wrapt=1.14.1=py310hca72f7f_0
wsproto=1.2.0=pypi_0
xz=5.4.6=h6c40b1e_0
yarl=1.9.4=pypi_0
zeromq=4.3.5=h93d8f39_0
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=h8a1eda9_5
google-cloud-bigquery
